Fix font rendering, refactor project structure
