+++ to-do-app/src/ToDoApp.css
@@ -7,7 +7,6 @@
   border: transparent;
   outline: none;
   background: transparent;
-  font-family: inherit;
   font-size: calc(10px + 2vmin);
   width: 100%;
   height: 100%;
@@ -104,7 +103,6 @@
   min-width: 296px;
   max-width: 780px;
   border-radius: 12px;
-  font-family: inherit;
   font-size: calc(15px + 2vmin);
 }
 
@@ -130,7 +128,6 @@
   border: transparent;
   outline: none;
   background: transparent;
-  font-family: inherit;
   text-align: center;
 }
 
@@ -141,7 +138,6 @@
   border-radius: 10px;
   margin-right: 10px;
   padding-right: 9px;
-  vertical-align: middle;
 }
 
 .To-Do-Space{
