+++ to-do-app/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import ToDoit from './ToDoit/index.js';
+import ToDoit from './components/ToDoit/index.js';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
